Fixed glfw throwing segmentation fault, requires reverting back to synchronous plotting (#112)
This PR enables the glfw window to be drawn when plot=True, but requires
reverting the implementation of plotter back to synchronous plotting.
This is because asynchronous plotting throws a segmentation fault within
glfw. Furthermore, every 'import glfw' statement must be preceded with
'import mujoco_py' to ensure that glfw is being imported properly.